Refactor test helpers and update unit tests for consistency
- Enhance helper functions in helpers.ts for better test coverage.
- Adjust unit tests in embedComposerContextService, embedIndexedDbService,
  embedSettingsService, embedStorageService, and parentClientAuthService
  for improved readability and maintainability.
