+++ src/test/unit/parentClientAuthService.test.ts
@@ -5,31 +5,8 @@ import {
 } from '../../lib/parentClientAuthService';
 import { STORAGE_KEYS } from '../../lib/constants';
 import type { ParentClientSessionData } from '../../lib/types';
-import { createMockConsole, type MockConsole, MockStorage } from '../helpers';
+import { createEmbedTestWindow, createMockConsole, type MockConsole, MockStorage } from '../helpers';
 import { EMBED_MESSAGE_NAMESPACE } from '../../lib/embedProtocol';
-
-function createMockWindow(search = '?parentOrigin=https%3A%2F%2Fparent.example.com') {
-    const listeners = new Map<string, (event: MessageEvent) => void>();
-    const parent = {
-        postMessage: vi.fn(),
-    };
-
-    const windowObj = {
-        self: {},
-        top: {},
-        parent,
-        location: { search },
-        document: { referrer: '' },
-        addEventListener: vi.fn((type: string, handler: (event: MessageEvent) => void) => {
-            listeners.set(type, handler);
-        }),
-        removeEventListener: vi.fn((type: string) => {
-            listeners.delete(type);
-        }),
-    } as unknown as Window;
-
-    return { windowObj, parent, listeners };
-}
 
 describe('ParentClientAuthService', () => {
     let mockConsole: MockConsole;
@@ -39,7 +16,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('auth.request/auth.result で接続し signer を生成する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const promise = service.connect({ capabilities: ['signEvent'] });
@@ -77,7 +54,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('rpc.request/rpc.result で signEvent を委譲する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const connectPromise = service.connect({ capabilities: ['signEvent'] });
@@ -124,7 +101,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('capabilities 未指定時は signEvent のみを既定要求する', () => {
-        const { windowObj, parent } = createMockWindow();
+        const { windowObj, parent } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         void service.connect({ timeoutMs: 1 }).catch(() => undefined);
@@ -179,7 +156,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('auth.error を受け取ると接続要求を明示的な認証エラーとして reject する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const promise = service.connect({ capabilities: ['signEvent'] });
@@ -204,7 +181,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('要求していない capability を含む auth.result を reject する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const promise = service.connect({ capabilities: ['signEvent'] });
@@ -228,7 +205,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('malformed な rpc.result を reject する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const connectPromise = service.connect({ capabilities: ['signEvent'] });
@@ -271,7 +248,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('requestId がない auth.result は warn して無視する', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
 
         const promise = service.connect({ capabilities: ['signEvent'] });
@@ -313,7 +290,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('auth.logout を受け取ると remote logout listener が呼ばれる', async () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
         const onRemoteLogout = vi.fn();
         service.onRemoteLogout(onRemoteLogout);
@@ -354,7 +331,7 @@ describe('ParentClientAuthService', () => {
     });
 
     it('auth.login を受け取ると remote login listener が呼ばれる', () => {
-        const { windowObj, parent, listeners } = createMockWindow();
+        const { windowObj, parent, listeners } = createEmbedTestWindow(undefined, { document: { referrer: '' } });
         const service = new ParentClientAuthService(windowObj, mockConsole);
         const onRemoteLogin = vi.fn();
         service.onRemoteLogin(onRemoteLogin);
